requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiohttp==3.9.1
aiofiles==23.2.1
jinja2==3.1.2
websockets==12.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
plotly==5.17.0
dash==2.16.1
dash-bootstrap-components==1.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
msal==1.25.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
simple-salesforce==1.12.5
slack-sdk==3.26.1
prometheus-client==0.19.0
structlog==23.2.0
rich==13.7.0
typer==0.9.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
